set xdebug.discover_client_host=1
Quoted from xdebug docs: https://xdebug.org/docs/step_debug

If PHP/Xdebug run on a different machine in the same sub-net, and you run your browser on the same host as your IDE, then you set the xdebug.discover_client_host setting to 1. Xdebug will then use the HTTP headers to find out the IP address of the host that initiated the debugging request, and use that IP address to connect to. This is a common way of set-up if you are sharing a development server among you and your team mates.